refactor bundle to new structure
