+++ src/Sulu/Bundle/PreviewBundle/Resources/config/config_preview.yml
@@ -6,7 +6,7 @@ framework:
 
 services:
     sulu.preview.voter:
-        class: Sulu\Bundle\PreviewBundle\Security\PreviewVoter
+        class: Sulu\Bundle\PreviewBundle\Infrastructure\Symfony\Security\PreviewVoter
         public: false
         tags:
             - { name: security.voter }
