--- src/Sulu/Bundle/PreviewBundle/Security/PreviewVoter.php
@@ -1,62 +1,0 @@
-<?php
-
-/*
- * This file is part of Sulu.
- *
- * (c) Sulu GmbH
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace Sulu\Bundle\PreviewBundle\Security;
-
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
-
-/**
- * A voter which disables security for the PreviewKernel.
- */
-class PreviewVoter implements VoterInterface
-{
-    /**
-     * Checks if the voter supports the given attribute.
-     *
-     * @param string $attribute An attribute
-     *
-     * @return bool true if this Voter supports the attribute, false otherwise
-     */
-    public function supportsAttribute($attribute)
-    {
-        return true;
-    }
-
-    /**
-     * Checks if the voter supports the given class.
-     *
-     * @param string $class A class name
-     *
-     * @return bool true if this Voter can process the class
-     */
-    public function supportsClass($class)
-    {
-        return true;
-    }
-
-    /**
-     * Returns the vote for the given parameters.
-     *
-     * This method must return one of the following constants:
-     * ACCESS_GRANTED, ACCESS_DENIED, or ACCESS_ABSTAIN.
-     *
-     * @param TokenInterface $token A TokenInterface instance
-     * @param object $object The object to secure
-     * @param array $attributes An array of attributes associated with the method being invoked
-     *
-     * @return int either ACCESS_GRANTED, ACCESS_ABSTAIN, or ACCESS_DENIED
-     */
-    public function vote(TokenInterface $token, $object, array $attributes)
-    {
-        return VoterInterface::ACCESS_GRANTED;
-    }
-}
